// backend-hedera/custodian/libs/shared/src/aef-report-management/aef-report-management.service.ts
import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
import { QueryRunner, Repository } from 'typeorm';
import * as fs from 'fs';
import * as path from 'path';
import * as ExcelJS from 'exceljs';
import * as moment from 'moment';
import { HelperService } from '../util/service/helper.service';
import { SerialNumberManagementService } from '../serial-number-management/service/serial-number-management.service';
import { InjectRepository } from '@nestjs/typeorm';
import { CreditTransactionsEntity } from '../carbon-credit-token/entity/credit.transfer.entity';
import { ConfigService } from '@nestjs/config';
import { AefActionsTableEntity } from './entity/aef.actions.table.entity';
import { FileHandlerInterface } from '../file-handler/filehandler.interface';
import { CreditBlocksEntity } from '../carbon-credit-token/entity/credit.blocks.entity';
import { CreditEventTypeEnum } from '../carbon-credit-token/enum/credit.event.type.enum';
import { plainToClass } from 'class-transformer';
import { AefActionTypeEnum } from './enum/aef.action.type.enum';
import { QueryDto } from '../util/dto/query.dto';
import { JWTPayload } from '../users/dto/jwt.payload.dto';
import { DataListResponseDto } from '../util/dto/data.list.response.dto';
import { OrganizationTypeEnum } from '../organization-type/enum/organization-type.enum';
import { RoleEnum } from '../role/enum/role.enum';
import { AefReportTypeEnum } from './enum/aef.report.type.enum';
import { ExportFileType } from './enum/export.file.type.enum';
import { DataExportHoldings } from './dto/data.export.holdings.dto';
import { DataExportActions } from './dto/data.export.actions.dto';
import { DataExportDto } from './dto/data.export.dto';
import { AefExportDto } from './dto/aef.export.dto';
import { CreditRetirementTypeEmnum } from '../carbon-credit-token/enum/credit.retirement.type.enum';
import { ProjectEntity } from '../project/entity/project.entity';

export const INF_SECTOR: { [key: string]: string } = {
    ENERGY: 'Energy',
    AGRICULTURE: 'Agriculture',
    HEALTH: 'Health',
    EDUCATION: 'Education',
    TRANSPORT: 'Transport',
    MANUFACTURING: 'Manufacturing',
    HOSPITALITY: 'Hospitality',
    FORESTRY: 'Forestry',
    WASTE: 'Waste',
    OTHER: 'Other',
};

export const INF_SECTORAL_SCOPE: { [key: string]: string } = {
    ENERGY_INDUSTRIES:
        'Energy Industries (Renewable – / Non-Renewable Sources) ',
    ENERGY_DISTRIBUTION: 'Energy Distribution',
    ENERGY_DEMAND: 'Energy Demand',
    AGRICULTURE: 'Agriculture',
    AFFORESTATION_AND_REFORESTATION: 'Afforestation and Reforestation',
    MANUFACTURING_INDUSTRIES: 'Manufacturing Industries',
    CHEMICAL_INDUSTRIES: 'Chemical Industries',
    METAL_PRODUCTION: 'Metal Production',
    TRANSPORT: 'Transport',
    WASTE_FROM_FUELS: 'Fugitive Emissions from Fuels (Solid, Oil and Gas) ',
    WASTE_HANDLING_AND_DISPOSAL: 'Waste Handling and Disposal',
    CONSTRUCTION: 'Construction',
    MINING_MINERAL_PRODUCTION: 'Mining/Mineral Production',
    FUGITIVE_EMISSIONS_PRODUCTION:
        'Fugitive Emissions from Production and Consumption of Halocarbons and Sulphur Hexafluoride',
    SOLVENT_USE: 'Solvent Use',
};
@Injectable()
export class AefReportManagementService {
    constructor(
        private readonly helperService: HelperService,
        private readonly serialNumberManagementService: SerialNumberManagementService,
        @InjectRepository(ProjectEntity)
        private projectRepository: Repository<ProjectEntity>,
        private readonly configService: ConfigService,
        @InjectRepository(AefActionsTableEntity)
        private aefActionsTableEntityRepository: Repository<AefActionsTableEntity>,
        private fileHandler: FileHandlerInterface,
    ) {}

    public async handleAefRecord(
        type: CreditEventTypeEnum,
        creditBlock: CreditBlocksEntity,
        queryRunner: QueryRunner,
        transaction?: CreditTransactionsEntity,
    ) {
        if (
            ![
                CreditEventTypeEnum.ISSUED,
                CreditEventTypeEnum.TRANSFERED,
                CreditEventTypeEnum.RETIRED,
            ].includes(type)
        ) {
            return;
        }
        const project = await this.projectRepository.findOne({
            where: { refId: creditBlock?.project?.refId },
        });
        const newAefActionRecord = plainToClass(AefActionsTableEntity, {
            creditBlockStartId:
                this.serialNumberManagementService.getBlockStartId(
                    creditBlock.serialNumber,
                ),
            creditBlockEndId: this.serialNumberManagementService.getBlockEndId(
                creditBlock.serialNumber,
            ),
            creditAmount: creditBlock.creditAmount,
            vintage: this.serialNumberManagementService.getVintage(
                creditBlock.serialNumber,
            ),
            sector: project.sector,
            sectoralScope: project.sectoralScope,
            projectAuthorizationTime: project.projectAuthorizationTime,
            authorizationId: project.authorizationId,
            actionTime: creditBlock.createdDate,
            aquiringParty: this.configService.get('AEF.defaultAquiringParty'),
        });
        if (type == CreditEventTypeEnum.ISSUED) {
            newAefActionRecord.actionType = AefActionTypeEnum.AUTHORIZATION;
        } else if (type == CreditEventTypeEnum.TRANSFERED) {
            newAefActionRecord.actionType = AefActionTypeEnum.TRANSFER;
        } else if (type == CreditEventTypeEnum.RETIRED) {
            if (
                transaction.retirementType ==
                CreditRetirementTypeEmnum.CROSS_BORDER_TRANSACTIONS
            ) {
                newAefActionRecord.actionType =
                    AefActionTypeEnum.CROSS_BOARDER_TRANSFER;
                newAefActionRecord.aquiringParty = transaction.country.alpha2;
            } else {
                newAefActionRecord.actionType = AefActionTypeEnum.RETIRE;
            }
            newAefActionRecord.creditAmount = transaction.creditAmount;
            newAefActionRecord.actionTime = transaction.createdDate;
        }

        await queryRunner.manager.save(newAefActionRecord);
    }

    public async queryAefRecords(
        query: QueryDto,
        abilityCondition: string,
        user: JWTPayload,
    ): Promise<DataListResponseDto> {
        if (
            user.organizationRole !=
                OrganizationTypeEnum.DESIGNATED_NATIONAL_AUTHORITY ||
            ![RoleEnum.Admin.toString(), RoleEnum.Root.toString()].includes(
                user.userRole,
            )
        ) {
            throw new HttpException(
                this.helperService.formatReqMessagesString(
                    'aef.unauthorized',
                    [],
                ),
                HttpStatus.UNAUTHORIZED,
            );
        }
        const resp = await this.aefActionsTableEntityRepository
            .createQueryBuilder('aefRecord')
            .where(this.helperService.generateWhereSQL(query, abilityCondition))
            .orderBy(
                query?.sort?.key && `"${query?.sort?.key}"`,
                query?.sort?.order,
                query?.sort?.nullFirst !== undefined
                    ? query?.sort?.nullFirst === true
                        ? 'NULLS FIRST'
                        : 'NULLS LAST'
                    : undefined,
            )
            .skip(query.size * query.page - query.size)
            .take(query.size)
            .getManyAndCount();
        const records = resp[0];
        const staticFields = this.configService.get('AEF');
        const updatedRecords = records.map((record) => ({
            ...record,
            ...staticFields,
            actionType: this.helperService.formatReqMessagesString(
                `aef.${record.actionType}`,
                [],
            ),
        }));
        return new DataListResponseDto(
            updatedRecords,
            resp.length > 1 ? resp[1] : undefined,
        );
    }

    async downloadAefReport(
        exportDto: AefExportDto,
        abilityCondition: string,
        user: JWTPayload,
    ) {
        if (
            user.organizationRole !=
                OrganizationTypeEnum.DESIGNATED_NATIONAL_AUTHORITY ||
            ![RoleEnum.Admin.toString(), RoleEnum.Root.toString()].includes(
                user.userRole,
            )
        ) {
            throw new HttpException(
                this.helperService.formatReqMessagesString(
                    'aef.unauthorized',
                    [],
                ),
                HttpStatus.UNAUTHORIZED,
            );
        }
        const query = new QueryDto();
        query.page = 1;
        query.size = await this.aefActionsTableEntityRepository.count();
        query.sort = {
            key: 'createdTime',
            order: 'DESC',
        };
        if (exportDto.reportType === AefReportTypeEnum.HOLDINGS) {
            query.filterAnd = [
                {
                    key: 'actionType',
                    operation: '=',
                    value: 'authorization',
                },
            ];
        }
        const resp = await this.queryAefRecords(query, abilityCondition, user);

        if (resp.total > 0) {
            let prepData;
            let localFileName;

            switch (exportDto.reportType) {
                case AefReportTypeEnum.HOLDINGS:
                    prepData = this.prepareHoldingsData(resp);
                    localFileName = `${AefReportTypeEnum.HOLDINGS}`;
                    break;
                case AefReportTypeEnum.ACTIONS:
                    prepData = this.prepareActionsData(resp);
                    localFileName = `${AefReportTypeEnum.ACTIONS}`;
                    break;
                default:
                    break;
            }

            const headers: string[] = [];
            const titleKeys = Object.keys(prepData[0]);
            for (const key of titleKeys) {
                headers.push(
                    this.helperService.formatReqMessagesString(
                        'aef.' + key,
                        [],
                    ),
                );
            }

            const path = await this.generateCsvOrExcel(
                prepData,
                exportDto.fileType === ExportFileType.CSV ? headers : titleKeys,
                this.helperService.formatReqMessagesString(localFileName, []),
                exportDto.reportType,
                exportDto.fileType,
            );

            return path;
        }
        throw new HttpException(
            this.helperService.formatReqMessagesString(
                'reportExport.nothingToExport',
                [],
            ),
            HttpStatus.BAD_REQUEST,
        );
    }

    private prepareHoldingsData(data: DataListResponseDto) {
        const exportData: DataExportHoldings[] = [];

        for (const report of data.data) {
            const dto: DataExportHoldings = new DataExportHoldings();
            dto.article6RecordId = report.artical6RecordId;
            dto.cooperativeApproach = report.cooperativeApproach;
            dto.firstUniqueIdentifier = report.firstUniqueIdentifier;
            dto.lastUniqueIdentifier = report.lastUniqueIdentifier;
            dto.creditBlockStartId = report.creditBlockStartId;
            dto.creditBlockEndId = report.creditBlockEndId;
            dto.metric = report.metric;
            dto.quantityInMetric = report.quantityInMetric;
            dto.creditAmount = report.creditAmount;
            dto.conversionFactor = report.conversionFactor;
            dto.firstTransferingParty = report.firstTransferingParty;
            dto.vintage = report.vintage;
            dto.sector = INF_SECTOR[report.sector];
            dto.sectoralScope = INF_SECTORAL_SCOPE[report.sectoralScope]
                ? INF_SECTORAL_SCOPE[report.sectoralScope]
                : 'NA';
            dto.projectAuthorizationTime = report.projectAuthorizationTime
                ? moment(parseInt(report.projectAuthorizationTime)).format(
                      'DD-MMM-YY',
                  )
                : '';
            dto.authorizationId = report.authorizationId;
            dto.purposeForAuthorization = report.purposeForAuthorization;
            dto.oimp = report.OIMP;
            dto.firstTransferDefinition = report.firstTransferDefinition;

            exportData.push(dto);
        }

        return exportData;
    }

    private prepareActionsData(data: DataListResponseDto) {
        const exportData: DataExportActions[] = [];

        for (const report of data.data) {
            const dto: DataExportActions = new DataExportActions();
            dto.article6RecordId = report.artical6RecordId;
            dto.cooperativeApproach = report.cooperativeApproach;
            dto.firstUniqueIdentifier = report.firstUniqueIdentifier;
            dto.lastUniqueIdentifier = report.lastUniqueIdentifier;
            dto.creditBlockStartId = report.creditBlockStartId;
            dto.creditBlockEndId = report.creditBlockEndId;
            dto.metric = report.metric;
            dto.quantityInMetric = report.quantityInMetric;
            dto.creditAmount = report.creditAmount;
            dto.conversionFactor = report.conversionFactor;
            dto.firstTransferingParty = report.firstTransferingParty;
            dto.vintage = report.vintage;
            dto.sector = INF_SECTOR[report.sector];
            dto.sectoralScope = INF_SECTORAL_SCOPE[report.sectoralScope]
                ? INF_SECTORAL_SCOPE[report.sectoralScope]
                : 'NA';
            dto.projectAuthorizationTime = report.projectAuthorizationTime
                ? moment(parseInt(report.projectAuthorizationTime)).format(
                      'DD-MMM-YY',
                  )
                : '';
            dto.authorizationId = report.authorizationId;
            dto.purposeForAuthorization = report.purposeForAuthorization;
            dto.oimp = report.OIMP;
            dto.firstTransferDefinition = report.firstTransferDefinition;
            dto.actionTime = report.actionTime
                ? moment(parseInt(report.actionTime)).format('DD-MMM-YY')
                : '';
            dto.actionType = report.actionType;
            dto.transferingParty = report.transferingParty;
            dto.aquiringParty = report.aquiringParty;
            dto.purposeForCancellation = report.purposeForCancellation;
            dto.actionBy = report.actionBy;
            dto.firstTransfer = report.firstTransferDefinition;

            exportData.push(dto);
        }

        return exportData;
    }

    async generateCsvOrExcel(
        data: DataExportDto[],
        headers: string[],
        fileName: string,
        reportType: AefReportTypeEnum,
        fileType: ExportFileType,
    ) {
        const currentDate = new Date();
        const year = currentDate.getFullYear();
        const month = (currentDate.getMonth() + 1).toString().padStart(2, '0');
        const day = currentDate.getDate().toString().padStart(2, '0');
        const hours = currentDate.getHours().toString().padStart(2, '0');
        const minutes = currentDate.getMinutes().toString().padStart(2, '0');
        const seconds = currentDate.getSeconds().toString().padStart(2, '0');

        const formattedDateTime = `${year}-${month}-${day}_${hours}-${minutes}-${seconds}`;
        const fileExtension = fileType === ExportFileType.CSV ? 'csv' : 'xlsx';
        const outputFileName = `${fileName}_${formattedDateTime}.${fileExtension}`;

        if (fileType === ExportFileType.CSV) {
            let csvContent = '';

            const refinedData = [];
            refinedData.push(headers);

            data.forEach((item) => {
                const values = Object.values(item).map((value) =>
                    value === undefined || value === null ? '' : value,
                );
                refinedData.push(values);
            });

            refinedData.forEach((row) => {
                const rowValues = row.map((value) => `"${value}"`).join(',');
                csvContent += rowValues + '\n';
            });

            fs.writeFileSync(outputFileName, csvContent);
        } else if (fileType === ExportFileType.XLSX) {
            switch (reportType) {
                case AefReportTypeEnum.ACTIONS:
                    await this.fillTemplate(
                        'aef_actions_template.xlsx',
                        'Actions',
                        headers,
                        data,
                        11,
                        outputFileName,
                    );
                    break;
                case AefReportTypeEnum.HOLDINGS:
                    await this.fillTemplate(
                        'aef_holdings_template.xlsx',
                        'Holdings',
                        headers,
                        data,
                        7,
                        outputFileName,
                    );
                    break;
                default:
                    break;
            }
        }

        const content = fs.readFileSync(outputFileName, { encoding: 'base64' });
        const url = await this.fileHandler.uploadFile(
            'exports/' + outputFileName,
            content,
        );

        console.log('Export completed', 'exports/', url);
        return { url, outputFileName };
    }

    async fillTemplate(
        templateName: string,
        sheetName: string,
        headers: string[],
        data: Record<string, any>[],
        startRow: number,
        outputFileName: string,
    ) {
        const templatePath = path.resolve(__dirname, 'templates', templateName);
        const wb = new ExcelJS.Workbook();
        await wb.xlsx.readFile(templatePath);

        const sheet = wb.getWorksheet(sheetName);
        if (!sheet)
            throw new Error(`Sheet ${sheetName} not found in ${templateName}`);

        if (templateName === 'aef_actions_template.xlsx') {
            sheet.getCell('B3').value = this.configService.get('AEF.party');
            sheet.getCell('B4').value = new Date().getFullYear();
            sheet.getCell('B3').font = { name: 'Times New Roman', size: 10 };
            sheet.getCell('B4').font = { name: 'Times New Roman', size: 10 };
        }

        let rowIdx = startRow;

        for (const item of data) {
            const row = sheet.getRow(rowIdx++);
            headers.forEach((key, colIdx) => {
                const cell = row.getCell(colIdx + 1);
                const v = item[key];
                cell.value = v == null ? '' : v;
                cell.font = { name: 'Times New Roman', size: 10 };
                cell.alignment = {
                    wrapText: true,
                    vertical: 'top',
                    horizontal: 'left',
                };
            });
            row.commit();
        }

        await wb.xlsx.writeFile(outputFileName);
        return outputFileName;
    }
}
